refactoring list challenges into InProgress
