+++ src/shared/ui/Skeletons/Skeletons.module.scss
@@ -2,7 +2,7 @@
 
 .wrapper {
     width: 100%;
-    height: 140px;
+    height: 120px;
     border-radius: 8px;
     padding: 15px;
     padding-bottom: 70px;
