+++ src/shared/ui/index.ts
@@ -4,3 +4,4 @@ export { Switcher } from './Switcher/Switcher';
 export { ErrorBoundary } from './Errors/ErrorBoundary';
 export { ErrorPage } from './Errors/ErrorPage';
 export { CardSkeleton } from './Skeletons/Skeletons';
+export { ListChallenges } from './ListChallenges/ListChallenges';
