+++ src/widgets/InProgressChallengesList/InProgressChallengesList.module.scss
@@ -1,0 +1,27 @@
+@use '@/shared/styles' as *;
+
+.container {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding-inline: 5px;
+    padding-top: 2vh;
+}
+
+.rowWrapper {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    gap: 40px;
+    margin-bottom: 1rem;
+}
+
+.title {
+    font-size: 24px;
+    font-weight: 400;
+    line-height: 29.88px;
+    text-align: center;
+}
