+++ src/widgets/index.ts
@@ -1,6 +1,6 @@
 export { Footer } from './Footer/Footer';
 export { Navigation } from './Navigation/ui/Navigation';
-export { ListChallenges } from './ListChallenges/ListChallenges';
+export { InProgressChallengesList } from './InProgressChallengesList/InProgressChallengesList';
 export { CreateForm } from './CreateForm/CreateForm';
 export { EditForm } from './EditForm/EditForm';
 export { ChallengeInfo } from './ChallengeInfo/ChallengeInfo';
